backend/requirements.txt
amqp==5.3.1
asgiref==3.11.0
billiard==4.2.4
celery==5.6.2
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
Django==6.0.1
django-cors-headers==4.9.0
django-environ==0.12.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
fonttools==4.61.1
kiwisolver==1.4.9
kombu==5.6.2
matplotlib==3.10.8
numpy==2.4.1
packaging==26.0
pandas==3.0.0
pillow==12.1.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.0
redis==7.1.0
requests==2.32.5
scipy==1.17.0
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
tzlocal==5.3.1
vine==5.1.0
wcwidth==0.5.0
gunicorn==22.0.0
dj-database-url==2.3.0